MACOSX port
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@2380 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip-ogl/src/GlBezierCurve.cpp
@@ -126,7 +126,7 @@ Coord GlBezierCurve::computeCurvePointOnCPU(float t) {
 		controlPoint[0] = controlPoints[i][0];
 		controlPoint[1] = controlPoints[i][1];
 		controlPoint[2] = controlPoints[i][2];
-		bezierPoint += controlPoint * pascalTriangle[controlPoints.size() - 1][i] * pow(t, i) * pow((1.0 - t), (controlPoints.size() - 1 - i));
+		bezierPoint += controlPoint * pascalTriangle[controlPoints.size() - 1][i] * pow((double) t, (double) i) * pow( (double) (1.0 - t), (double) (controlPoints.size() - 1 - i));
 	}
 	return Coord(bezierPoint[0], bezierPoint[1], bezierPoint[2]);
 }
